WIP: Start sketching in `ProjectFindView`
