+++ crates/find/src/project_find.rs
@@ -1,7 +1,24 @@
-use crate::SearchMode;
-use editor::MultiBuffer;
-use gpui::{Entity, ModelContext, ModelHandle, Task};
+use anyhow::Result;
+use editor::{Editor, MultiBuffer};
+use gpui::{
+    action, elements::*, keymap::Binding, ElementBox, Entity, Handle, ModelContext, ModelHandle,
+    MutableAppContext, Task, View, ViewContext, ViewHandle,
+};
 use project::Project;
+use std::{borrow::Borrow, sync::Arc};
+use workspace::Workspace;
+
+action!(Deploy);
+action!(Search);
+
+pub fn init(cx: &mut MutableAppContext) {
+    cx.add_bindings([
+        Binding::new("cmd-shift-f", Deploy, None),
+        Binding::new("enter", Search, Some("ProjectFindView")),
+    ]);
+    cx.add_action(ProjectFindView::deploy);
+    cx.add_async_action(ProjectFindView::search);
+}
 
 struct ProjectFind {
     last_search: SearchParams,
@@ -20,6 +37,8 @@ struct SearchParams {
 
 struct ProjectFindView {
     model: ModelHandle<ProjectFind>,
+    query_editor: ViewHandle<Editor>,
+    results_editor: ViewHandle<Editor>,
 }
 
 impl Entity for ProjectFind {
@@ -42,5 +61,104 @@ impl ProjectFind {
             //
             None
         });
+    }
+}
+
+impl workspace::Item for ProjectFind {
+    type View = ProjectFindView;
+
+    fn build_view(
+        model: ModelHandle<Self>,
+        workspace: &workspace::Workspace,
+        nav_history: workspace::ItemNavHistory,
+        cx: &mut gpui::ViewContext<Self::View>,
+    ) -> Self::View {
+        let settings = workspace.settings();
+        let excerpts = model.read(cx).excerpts.clone();
+        let build_settings = editor::settings_builder(excerpts.downgrade(), workspace.settings());
+        ProjectFindView {
+            model,
+            query_editor: cx.add_view(|cx| Editor::single_line(build_settings.clone(), cx)),
+            results_editor: cx.add_view(|cx| {
+                Editor::for_buffer(
+                    excerpts,
+                    build_settings,
+                    Some(workspace.project().clone()),
+                    cx,
+                )
+            }),
+        }
+    }
+
+    fn project_path(&self) -> Option<project::ProjectPath> {
+        None
+    }
+}
+
+impl Entity for ProjectFindView {
+    type Event = ();
+}
+
+impl View for ProjectFindView {
+    fn ui_name() -> &'static str {
+        "ProjectFindView"
+    }
+
+    fn render(&mut self, cx: &mut gpui::RenderContext<'_, Self>) -> ElementBox {
+        Flex::column()
+            .with_child(ChildView::new(&self.query_editor).boxed())
+            .with_child(ChildView::new(&self.results_editor).boxed())
+            .boxed()
+    }
+}
+
+impl workspace::ItemView for ProjectFindView {
+    fn item_id(&self, cx: &gpui::AppContext) -> usize {
+        self.model.id()
+    }
+
+    fn tab_content(&self, style: &theme::Tab, cx: &gpui::AppContext) -> ElementBox {
+        Label::new("Project Find".to_string(), style.label.clone()).boxed()
+    }
+
+    fn project_path(&self, cx: &gpui::AppContext) -> Option<project::ProjectPath> {
+        None
+    }
+
+    fn can_save(&self, _: &gpui::AppContext) -> bool {
+        true
+    }
+
+    fn save(
+        &mut self,
+        project: ModelHandle<Project>,
+        cx: &mut ViewContext<Self>,
+    ) -> Task<anyhow::Result<()>> {
+        self.results_editor
+            .update(cx, |editor, cx| editor.save(project, cx))
+    }
+
+    fn can_save_as(&self, cx: &gpui::AppContext) -> bool {
+        false
+    }
+
+    fn save_as(
+        &mut self,
+        project: ModelHandle<Project>,
+        abs_path: std::path::PathBuf,
+        cx: &mut ViewContext<Self>,
+    ) -> Task<anyhow::Result<()>> {
+        unreachable!("save_as should not have been called")
+    }
+}
+
+impl ProjectFindView {
+    fn deploy(workspace: &mut Workspace, _: &Deploy, cx: &mut ViewContext<Workspace>) {
+        let model = cx.add_model(|cx| ProjectFind::new(workspace.project().clone(), cx));
+        workspace.open_item(model, cx);
+    }
+
+    fn search(&mut self, _: &Search, cx: &mut ViewContext<Self>) -> Option<Task<Result<()>>> {
+        todo!()
     }
 }
